+++ src/api/c++/include/sqlrelay/sqlrclient.h
@@ -345,23 +345,22 @@ class sqlrcursor {
 				// Returns the type of the specified column.
 		int	getColumnLength(int col);
 		int	getColumnLength(const char *col);
-				// Returns the length of the specified column.
-				// (This may be different than precision because
-				// some databases return the number of bytes
-				// required to store the data for numeric types
-				// rather than number of digits)
-		unsigned short	getColumnPrecision(int col);
-		unsigned short	getColumnPrecision(const char *col);
+				// Returns the number of bytes required on
+				// the server to store the data.
+		unsigned long	getColumnPrecision(int col);
+		unsigned long	getColumnPrecision(const char *col);
 				// Returns the precision of the specified
 				// column.
-				// (Precision is the total number of digits in
-				// a number.  eg: 123.45 has a precision of 5)
-		unsigned short	getColumnScale(int col);
-		unsigned short	getColumnScale(const char *col);
+				// Precision is the total number of digits in
+				// a number.  eg: 123.45 has a precision of 5.
+				// For non-numeric types, it's the number of
+				// characters in the string.
+		unsigned long	getColumnScale(int col);
+		unsigned long	getColumnScale(const char *col);
 				// Returns the scale of the specified column.
-				// (Scale is the total number of digits to the
+				// Scale is the total number of digits to the
 				// right of the decimal point in a number.
-				// eg: 123.45 has a scale of 2)
+				// eg: 123.45 has a scale of 2.
 		int	getLongest(int col);
 		int	getLongest(const char *col);
 				// Returns the length of the longest field
